requirements.txt
Flask==2.3.2
python-dotenv==1.0.1
openai==1.23.2
httpx==0.27.2
gunicorn==21.2.0
Flask-Compress==1.14
Flask-Caching==2.1.0
gevent==23.9.1
